Add error diagnostic for integer literals that don't fit into `uint64_t` (#9208)
UndefinedBehaviorSanitizer reported a signed integer overflow in
`getIntegerLiteralValue`. The root cause is that the
`IntegerLiteralValue` type is an alias for `int64_t` and the Slang
compiler doesn't check for overflows when parsing integer literals into
this type.

As a result, in the following example, Slang lets the value of `x`
silently overflow to 0:

```slang
int x = 18446744073709551616;
```

For comparison, by default, Clang emits an error:

```
a.c:1:9: error: integer literal is too large to be represented in any integer type
    1 | int x = 18446744073709551616;
      |         ^
1 error generated.
```

And GCC emits a warning:

```
a.c:1:9: warning: integer constant is too large for its type
    1 | int x = 18446744073709551616;
      |         ^~~~~~~~~~~~~~~~~~~~
```

This PR follows Clang by adding an error diagnostic instead of a
warning.

Since the lexer scans the minus sign of negative numbers as a separate
token, an integer literal is always a positive number. When lexing the
value of `INT64_MIN`, the integer literal is thus `INT64_MAX + 1`, which
overflows when stored into an `IntegerLiteralValue`/`int64_t` despite
being valid. Therefore, this PR changes `getIntegerLiteralValue` to use
`uint64_t` internally.

Related to #9099.

---------

Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>