Fix UBSan errors in `slang-memory-arena.h`
`MemoryArena` behaves similarly to C++ standard library containers:
memory is only allocated once one of the allocation functions is called
(the constructor doesn't allocate a block in advance).

Its allocation functions perform pointer arithmetic to check whether
there is enough free space left in the current block or if a new block
needs to be allocated. However, applying a non-zero offset to a null
pointer is undefined behavior. This PR works around this by changing
these allocation functions to return early if the pointer to the current
block is null.

Related to #9099.