Add SLANG_ASSERT environment variable support (#9313)
This commit replaces all of the system `assert()` with SLANG_ASSERT
macro so that all of assertion is handled in one place consistently in
`handleSignal` function.

This commit also allows us to change the behavior of the assertion at
runtime via a new environment variable "SLANG_ASSERT".

When it is set to "system", it will call the system `assert()`. On
Windows, it will open a popup dialog window and wait for us to attach a
debuger to the process.

When it is set to "debugbreak", it will call `__debugbreak()`, which can
be handy when you are debugging with a debugger.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>