slangd: Canonicalize all paths found from search paths into URIs (#9279)
Fixes #8286

Many uses of `URI::fromLocalFilePath` in the Slang language server are
on paths that are canonicalized (made absolute) by `Path::getCanonical`,
but currently `LanguageServerCore::gotoDefinition` does not canonicalize
paths from `PathInfo::foundPath` before transforming them into URIs with
`fromLocalFilePath` and returning.
As a result, if `foundPath` is relative, which will be the case if there
is a relative path in the `slang.additionalSearchPaths` setting,
`gotoDefinition` returns a the relative path prepended with `file://`,
which is not a valid URI for a local filepath according to the [IETF URI
Spec](https://datatracker.ietf.org/doc/html/rfc3986).

This change adds calls to `getCanonical` on every path used by
`fromLocalFilePath` for which we do not already use `getCanonical` on.