Add missing DebugLines in two cases
In SPIR-V legalization, we insert blocks into IRIfElse insts to avoid
the situation where either the 'true' or 'false' block is the same as
the 'after' block. To accomplish this, we produce a trivial block that
just jumps to the after block. This produces a problem in the SPIR-V
output, where we have a block without a DebugLine entry, which causes
[whichever compiler consumes the SPIR-V] to attribute that block to line
0.

To fix this first problem, find the first IRDebugLine inst in the
'after' block and copy it to the new block we're creating.

A similar issue occurs during constructSSA: critical edge elimination
can create new blocks which lack a DebugLine entry. Do the same thing
there.

Some code in slang-ir-simplify-cfg.cpp assumes that DebugLines won't be
present in trivial blocks, so edit those to skip over DebugLines when
searching for IRUnconditonalBranch instructions.

The second case is in visitLogicOperatorShortCircuitExpr: this function
produces three blocks without debugLines, which causes the same problem
as above. Simply emit debugline info in these blocks.

Fixes #9322.