Get target-specific layout for DescriptorHandle<T>, refactor Type-kind mapping in layout creation (#8383)
Fixes https://github.com/shader-slang/slang/issues/6657

This implements a function `GetDescriptorHandleLayout` to
`SimpleLayoutRulesImpl` with target-specific implementations, and uses
it in `getSimpleVaryingParameterTypeLayout` to return the appropriate
layout for a varying `DescriptorHandle<T>` parameter on a given target,
which allows varying `DescriptorHandle` inputs to be assigned the
appropriate bindings.
For bindless targets (e.g. CPU, Metal, CUDA),
`GetDescriptorHandleLayout` returns the layout of `T`. For SPIR-V (and
GLSL), it returns a `uint64_t` layout if `spvBindlessTextureNV` is
enabled, otherwise it returns a `uint2` layout. For other non-bindless
targets (e.g. HLSL), it returns a `uint2` layout.

The type layout returned for `DescriptorHandle` retains the
`DescriptorHandle` type in reflection, instead of the underlying type's
name.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>