Add support for UndefinedBehaviorSanitizer
Changes the `SLANG_ENABLE_ASAN` CMake option to also enable
UndefinedBehaviorSanitizer (UBSan) if the C++ compiler is Clang or GCC
(MSVC only supports ASan, not UBSan).

This PR only makes the changes necessary for building Slang with UBSan.
As there are many UBSan errors to fix, they will be fixed by other PRs
(some of which have already been merged).

However, one of UBSan's type checks was causing a segfault/crash as it
was dereferencing a null `this` pointer. The root cause is that
`getCurrentASTBuilder()` was returning a null pointer in some test
cases. To fix this, global session initialization now sets the session's
root AST builder as the current AST builder, restoring it on
destruction. This also fixes some of the UBSan errors (member function
calls on null `ASTBuilder` pointers).

Fixes #9099.