Add "release-asserts-only" option to SLANG_ASSERT environment variable (#9363)
When the environment variable `SLANG_ASSERT` is set to
"release-asserts-only" on Windows, any assertion failures will be
ignored. This behavior will be same to how Release build works and it
can help us to reproduce issues that shows different behavior between
Release and Debug builds.

A new function `handleAssert()` had to be introduced, because the
pre-existing `handleSignal()` has [[noreturn]] which cannot work if we
want to just return and resume the normal flow.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>
Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>