Prevent reads from mesh shader outputs
Reading from mesh shader outputs isn't allowed in HLSL:
    https://microsoft.github.io/DirectX-Specs/d3d/MeshShader.html#shared-output-arrays

... or in SPIR-V:

VUID-StandaloneSpirv-MeshEXT-07107
In mesh shaders using the MeshEXT Execution Model all variables declared
in the Output Storage Class must not be read

Add a validation pass when generating IR to check for loads
from IRMeshOutputRef, and error if one is found.

Also, in parameter passing modes that require use of a temporary, detect
when we create a load from a write-only value, and yield an error.

Fixes #9315