Add missing DebugLines in two cases (#9371)
When emitting SPIR-V, we can't have 'direct' merges -- that is, 'then',
'else', and 'after' must all be distinct. Legalization does this, but it
produces a block without any debug location information, which (when
that SPIR-V is fed into another compiler) leaves that block without a
valid DebugLine.

To fix this first problem, emit an empty 'else' block with a debugline
when emitting SPIR-V.

The second case is in visitLogicOperatorShortCircuitExpr: this function
produces three blocks without debugLines, which causes the same problem
as above. Simply emit debugline info in these blocks.

Fixes #9322.

---------

Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>
Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>