Re-write the logic to get coopMat tile size
Slang doesn't support constexpr for a function, so we cannot model
the shape selection mechansim in a function, instead we have to
module it in type traits as well. However, we don't support
target_switch at struct scope. Therefore, the target has to be a
generic paramter, and call site will use target_switch to provide
it's cuda or spirv target as an enum to the struct.