LLVM IR target (#8960)
# Direct-to-LLVM IR target

**TL;DR**: This PR implements a new LLVM IR target which does not go
through C++ as CPU targets have previously done. It translates Slang IR
directly into LLVM IR.

This target is documented in `docs/llvm-target.md`, so I am not
repeating the details here.

Benefits from the user perspective:
* Potentially faster compile times (by skipping the extra step of C++)
* Support for arbitrary type layouts (e.g. scalar layout & std430)
* Better debug info, allowing native-feeling debugging of Slang code
* Potential performance improvements through ability to communicate
Slang semantics as LLVM attributes (e.g. `noalias` for `out`/`inout`
params)

Benefits from a compiler developer perspective:
* No reliance on a downstream compiler being available on the system
(like Clang, G++ or MSVC)
* Slang-on-CPU semantics are no longer bound by what is expressible in
C++
* Portable vectorization & FP16 support via LLVM IR intrinsics
* Allows us to take advantage of various LLVM passes ([like
DeSPMD](#despmd-integration))
* The _general vibe_ of Slang no longer being a "mere" transpiler on CPU

## Brief code overview

The LLVM target is split in two parts: *Emitter*
(`source/slang/slang-emit-llvm.cpp`) and *Builder*
(`source/slang-llvm/slang-llvm-builder.cpp`).

This split is done because we practically cannot access both Slang IR
and LLVM IR in the same source file: LLVM is fully optional in Slang, so
`slang-compiler.dll` cannot link to it and therefore cannot work with
LLVM APIs such as their IRBuilder. Conversely, `slang-llvm.dll` should
not link to `slang-compiler.dll` in order to avoid cyclical linking and
issues with debug/release build compatibility on Windows.

### Emitter

The *Emitter* side works similarly to the other `slang-emit-*.cpp`
files. It handles Slang IR and constructs LLVM IR using the *Builder*.

The main class `LLVMEmitter` handles traversing Slang IR and emitting
instructions. The central function here is `emitLLVMInstruction()`.

`LLVMTypeTranslator` handles types and their layout rules. `getType()`
returns the normal LLVM types, and `getDebugType()` gets debug versions
with far more metadata.

### Builder

The *Builder* side exposes a COM interface, `ILLVMBuilder`, which can be
used to construct LLVM IR. It does a similar job as `llvm::IRBuilder`,
but is streamlined to make the interface smaller. It also handles code
generation. It gives out LLVM instructions and types to the *Emitter* as
opaque pointers.

Internally, the builder uses LLVM's IRBuilder and DIBuilder interfaces.
An alternate approach would've been to just generate the LLVM IR in text
form by ourselves; however, that is a brittle approach when updating
LLVM versions. By using LLVM's APIs, we're much more likely to get
compile errors when LLVM IR changes with some update, making this
approach more maintainable.

Architecturally, the *Builder* is quite straightforward, most functions
just emit one LLVM instruction, although they often smooth over corner
cases and inconsistencies in LLVM's API. The more interesting functions
are the ones doing codegen:

* `emitInlineIRFunction()`
* `emitComputeEntryPointWorkGroup()`
* `emitComputeEntryPointDispatcher()`
* `generateAssembly()`
* `generateObjectCode()`
* `generateJITLibrary()`

## Meta discussion

I have worked on this PR for a long while now, but I'm not particularly
attached to any architectural choices here. I'm open for any
suggestions, I'll rather fix any worries than rush this PR. I'm fine
with large refactors.

I do realize that this is a massive PR and a lot to chew on for
potential reviewers. Let me know if you'd like me to rather split this
up into separate PRs, e.g. as follows:

1. Builder addition
2. Emitter addition
3. LLVM intrinsics in `*.meta.slang`
4. test enablement

I chose to begin with this big all-in-one PR because it's testable as a
whole, whereas partial implementations would not be able to run tests.

## Future work & plans

This section is "optional" reading, it doesn't affect this PR directly.

These are items that are left unimplemented in this PR to limit the
scope. Some of these are long-term "wishlist" features, which become
much more attainable with the direct-to-LLVM emitter.

These are in a rough order of importance _to me_. I'm planning to work
on some of this stuff myself. If someone else wants to do any of the
things listed below, that's also cool by me, I can review and provide
help with that.

### Texture and acceleration structure support

The problem with these types is that there's no "objective" or even
"de-facto" memory layout for them on CPU:
- Textures: Z-order curve or not? NPOT textures? Mipmaps?
- TLAS: Embree? Some BVH from a physics library the user has already
integrated in their game?

Especially the latter is worrying to me, since one known use-case for
the CPU backends of Slang is particle simulation (#8244), where one
could want to do collision testing with the scene using an AS. It would
be good to allow users to take advantage of the acceleration structures
of their physics engines.

I think these features could be implemented by lowering their types as
opaque pointers, and forward declaring all sampling / tracing functions
when they are used. It is then up to the user to link proper
implementations for the operations they use. `slang-rt` could perhaps
provide some default implementations.

### Binding generator

Using arbitrary C libraries from the LLVM target requires generating
some kind of Slang bindings for the libraries. Getting the correct
calling convention for passing structure parameters is challenging to do
portably, and effectively requires interaction with Clang.

I've already made a [binding
generator](https://github.com/juliusikkala/slang-cpu-utils/blob/main/bindgen/bindgen.py)
for the C++-based CPU target, but its dependency on the Python bindings
of Clang is limiting and I'll probably rewrite this in C++.

I'm planning to work on this in my own repo first. I may propose it for
inclusion in `tools/` or `extras/` if it is deemed to be of general
interest.

### Pointer size refactor

Just get rid of `SLANG_PTR_IS_32` & `SLANG_PTR_IS_64` defining the
_target's_ pointer size all around Slang. When the target is LLVM, we
should check the pointer size based on the target triple.

### DeSPMD integration

This is an LLVM IR pass that turns a function into a thread group /
workgroup, taking barriers and subgroup operations properly into
account. It has been discussed in #8244 and would help address that
issue.

I have been privately given access to it, so I can begin working on this
quite soon, but I cannot release code until related ongoing research has
been published.

### LLVM layout

It would be great if the "LLVM type layout" _really_ used the DataLayout
of LLVM's target machine to determine alignments. This may require quite
a bit refactoring all around Slang to work, so I didn't tackle this one
yet. It could also feel unpredictable to users that the vector layout
can depend on the platform, so I'm not sure whether this is actually
desirable. The layout in this PR plays it safe; things may be
excessively aligned for some platforms.

### Existing C++-specific features

`class` and COM interfaces are not yet supported on the LLVM targets.

### GPU targets (OpenCL SPIR-V, DXIL)

This new LLVM target could provide a path to address #2533. Earlier,
I've taken a brief stab at modifying the existing SPIR-V emitter to
allow for OpenCL support, but there are major differences making it more
effort than I was willing to expend at that time.

LLVM has targets for OpenCL SPIR-V and DXIL. I'm not yet sure of all the
things that would be necessary for supporting these, but it would at
least require mapping Slang resource types to these targets somehow and
handling pointer address spaces correctly.

---------

Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>