Fix static-const lowering for interface generic methods
This commit fixes the incorrect hoisting of specialize inst when
its operands uses both the generaic-type-params and the static const
requirements in interface generics.

As a solution, when lowering the outer generic for anything inside
a generic interface that represent the generic parameters of the
interface, always create a fresh ThisType and ThisTypeWitness.