Fix incorrect use of reinterpret_cast and union (#9458)
I wasn't entirely sure, according to discussion
https://github.com/llvm/llvm-project/issues/173167, it turned out that
compilers do not provide type-punning in such cases and such use of
reinterpret_cast and union leads to undefined behavior.

```cpp
struct SlangUUID
    {
        uint32_t data1;
        uint16_t data2;
        uint16_t data3;
        uint8_t data4[8];
    };

bool foo(const SlangUUID& aIn, const SlangUUID& bIn)
{
    typedef uint32_t CmpType;
    union GuidCompare
    {
        SlangUUID guid;
        CmpType data[sizeof(SlangUUID) / sizeof(CmpType)];
    };
    const CmpType* a = reinterpret_cast<const GuidCompare&>(aIn).data;
    const CmpType* b = reinterpret_cast<const GuidCompare&>(bIn).data;

    return ((a[0] ^ b[0]) | (a[1] ^ b[1]) | (a[2] ^ b[2]) | (a[3] ^ b[3])) == 0;
}
```
the _structure_ is converted to a _union_ (with a member of array of
`uint32[4]`), and these types are not pointer-interconvertible (this is
undefined behavior according to the C++ standard)
(**[basic.compound#5]**).
https://eel.is/c++draft/basic.compound#5
https://eel.is/c++draft/conv.qual#2
https://eel.is/c++draft/class.mem#general-30