Fix scope of local variables from inlined functions (#9418)
DebugScope lookup in the SPIR-V backend simply searches up the IR tree
for the first IRFunc or IRModuleInst that maps to a valid lexical scope.

This causes scope lookup for (for example) local variables from inlined
functions to be attributed to the scope of the 'caller/inliner'
function, rather than the 'callee/inlined' function.

To fix this, in findDebugScope, search backwards in IRBlocks for a
DebugScope instruction, and use that instead.

This also fixes a bug in forceinline-multiple-blocks.slang: the test
expected the DebugInlinedAt instruction to appear six times, even though
the program *should* inline nine times:
- main calls inlineMultipleBasicBlocks twice (2), each of which call
calculateAdjustment three times, for a total of 6 calculateAdjustment
calls.
- main calls calculateAdjustment once

Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>