Fix mishandling of induction values in loop unrolling.
The current unrolling logic assumes that a loop induction value
(a IRParam in a loop's target block) is never used outside the loop body.

However, there are cases where this assumption does not hold. The unrolling
logic would create incorrect code in this case.

Specifically, given:
```
int calcOffset(uint rank)
{
    uint offset = 0;
    [ForceUnroll]
    for (int i = 0; i < 3; i++)
    {
        if (i >= 2)
            break;
        offset += i;
    }
    return offset;
}
```

The induction value `offset` is used after the break block of the loop.
The current unrolling logic works by "peeling" off one iteration from the loop,
then modify the loop structure to start from the next iteration. After this transformation,
the `IRParam` in the remaining loop target block is no longer dominating the outsider
use that the original `IRParam` used to, and this leads to incorrect code.

This fix works by detecting such out of loop uses of induction values, and
insert a new `IRParam` in the break block to represent the induction value
at loop exit, and replace the outsider uses with this new `IRParam`. This way
the existing unrolling logic will be correct and can remain simple.