Reduce copying inside SourceFile::setContents; fix non-BOM UTF-16 offset and skip UTF-16 checks if file size is odd (#9407)
Previously, `SourceFioe::setContents(ISlangBlob*)` would always copy the
contents twice: once inside `CharEncoding::decode`, and once when
creating a `RawBlob`.

This commit adds a fast-path for the common case of ASCII or UTF-8
without a BOM, in which case we don't need to copy at all. In the slow
path we can also save a copy by moving the list's buffer into the
`RawBlob`. The construction for this move is kind of awkward though --
would it be OK if I added a `RawBlob` `createMove` function from a
`List`, or maybe templated `ListBlob` so that it can handle signed
`char`s?

Note that this isn't quite optimal for `SourceFile::setContents(const
String&)` when the input is UTF-16. The `const String&` version always
copies at least once, which is unnecessary when we need to copy on the
decode.

This also includes a quick fix for `CharEncoding::determineEncoding`: if
it determines the file is UTF-16 without a BOM, it reported an offset of
2 when it should have reported an offset of 0, so the first UTF-16
character would be skipped. This bug can be reproduced by running
`slangc -target glsl` on a file like this:
[utf16.slang.txt](https://github.com/user-attachments/files/24234749/utf16.slang.txt).
Even though the file starts with `#version 460`, Slang ignored the first
character and emitted

```
utf16.slang(1): error 20001: unexpected integer literal, expected identifier
version 460
        ^~~
```

Additionally, we can skip the nontrivial (5 instructions/byte) search
for UTF-16 patterns half the time; if the file length is odd, it can't
be UTF-16. (Although maybe since the non-BOM UTF-16 path was subtly
broken, the wiser move would be to remove it altogether so determining
file encodings goes from linear-time to constant-time?)

Thanks!

---------

Signed-off-by: Nia Bickford <nbickford@nvidia.com>
Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>