Fix mishandling of induction values in loop unrolling. (#9476)
The current unrolling logic assumes that a loop induction value (a
IRParam in a loop's target block) is never used outside the loop body.

However, there are cases where this assumption does not hold. The
unrolling logic would create incorrect code in this case.

Specifically, given:
```
int calcOffset(uint rank)
{
    uint offset = 0;
    [ForceUnroll]
    for (int i = 0; i < 3; i++)
    {
        if (i >= 2)
            break;
        offset += i;
    }
    return offset;
}
```

The induction value `offset` is used after the break block of the loop.
The current unrolling logic works by "peeling" off one iteration from
the loop, then modify the loop structure to start from the next
iteration. After this transformation, the `IRParam` in the remaining
loop target block is no longer dominating the outsider use that the
original `IRParam` used to, and this leads to incorrect code.

This fix works by detecting such out of loop uses of induction values,
and insert a new `IRVar` before the loop to store the updated induction
value (updated every iteration), and replace all out of loop uses of the
induction value with a load of the Var.

Another issue fixed by this PR is that during Multi-level break
legalization, if we are about to make available an pointer-access-chain
instruction defined in some inner scope to the outer scope, don't store
the result of that access chain (which is a pointer type) in a variable
as this leads to incorrect/illegal variable pointer usage. Instead,
duplicate the access chain in the use sites.

Closes #9456.