Preserve emtpy tuple elements in `createGLSLGlobalVaryings`. (#9479)
Closes #9459.

The root cause of the crash is that if one of the struct fields is empty
(`void`), `createGLSLGlobalVarying` will return a TupleVal that does not
have a corresponding element for the empty field, but `getSubscriptVal`
and many other places in the legalize GLSL varyings pass expects that a
TupleVal lowered from a struct type to always have the same number of
entries as the struct fields. So the fix is to always insert a `none`
element for empty struct fields while legalizing a struct type to
`ScalarizedVal`.