Activtely synthesize `AttributeDecl` for user defined attribute structs.
This change fixes a nuanced bug around precompiled modules and user defined attributes:

Given module `test` that imports module `foo`, where `foo` defines a struct type with
[__AttributeUsage] to make it a user defined attribute, but `foo` itself does not make
any use of that attribute.
Then `test` contains decls using the user defined attribute from `foo`.
Now, if we precompile `test` into a module, but leave `foo` as source, and then try
to load the precompiled `test` module, we had a bug where the user defined attribute
cannot be found from the imported `foo` module, leading to an error.

The reason this was happening is that for every struct decl with [__AttributeUsage], we
will synthesize a `AttributeDecl` representing the attribute, and the AttributeDecl, instead
of the defining struct type decl is what gets referenced from any use sites of the attribute.
However, the synthesis of the AttributeDecl was happening on-demand when we lookup an attribute
from the [] bracket syntax on use sites. Therefore, when we serialize `foo` itself, since it
does not use the attribute itself, the AttributeDecl was never synthesized and serialized.
Now when we load the precompiled `test` module, it expects this AttributeDecl to be present in
`foo`, but it is not, leading to the error.

This issue has been fixed by ensuring that AttributeDecls are always synthesized actively upon
checking of [__AttributeUsage] on struct decls, regardless of whether the attribute
is used in the defining module or not.

Closes #8880